剔除无效文件
