+++ example/Example 4 - Queue/stm/task1.py
@@ -1,0 +1,6 @@
+import time
+from . import app
+@app.task(ignore_result=True)
+def add(x, y):
+    time.sleep(2)
+    return x + y
