+++ example/old_version/common.json
@@ -1,0 +1,17 @@
+{
+	"risk": {
+		"conn": "mysql+mysqlconnector://quant:AUsYCJ4cMa@miglab.jios.org:3306/quant",
+		"model": "model",
+		"table": "RiskExposure",
+		"trade_date": "trade_date"
+	},
+	"market": {
+		"conn": "mysql+mysqlconnector://quant:AUsYCJ4cMa@miglab.jios.org:3306/quant",
+		"model": "model",
+		"table": "Market",
+		"trade_date": "trade_date",
+		"code": "code",
+		"tclose": "closePrice",
+		"columns": []
+	}
+}
