+++ example/old_version/q7_task/celery1.pid
@@ -1,0 +1,1 @@
+2392
